requirements.txt
colorama==0.4.6
langchain==0.1.19
langchain_experimental==0.0.58
langchain_openai==0.1.6
pandas==2.2.1
pydantic==1.10.12
python-dotenv==1.0.1
chroma-hnswlib==0.7.2
chromadb==0.4.5
tiktoken==0.7.0
requests~=2.32.3
openai~=1.35.7